feat: Enable bubble content in right column panel (#2522)

+++ app/script/view_model/panel/ParticipantsViewModel.js
@@ -20,9 +20,10 @@
 'use strict';
 
 window.z = window.z || {};
-window.z.ViewModel = z.ViewModel || {};
+window.z.viewModel = z.viewModel || {};
+window.z.viewModel.panel = z.viewModel.panel || {};
 
-z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
+z.viewModel.panel.ParticipantsViewModel = class ParticipantsViewModel {
   static get CONFIG() {
     return {
       ADD_STATES: [ParticipantsViewModel.STATE.ADD_PEOPLE, ParticipantsViewModel.STATE.ADD_SERVICE],
@@ -40,7 +41,7 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     };
   }
 
-  constructor(elementId, conversationRepository, integrationRepository, teamRepository, userRepository) {
+  constructor(mainViewModel, panelViewModel, repositories) {
     this.clickOnAddPeople = this.clickOnAddPeople.bind(this);
     this.clickOnPending = this.clickOnPending.bind(this);
     this.clickOnMemberBack = this.clickOnMemberBack.bind(this);
@@ -48,37 +49,38 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     this.clickOnShowParticipant = this.clickOnShowParticipant.bind(this);
     this.clickToAddService = this.clickToAddService.bind(this);
     this.clickToBlock = this.clickToBlock.bind(this);
-    this.clickToConnect = this.clickToConnect.bind(this);
     this.clickToLeave = this.clickToLeave.bind(this);
     this.clickToRemoveMember = this.clickToRemoveMember.bind(this);
     this.clickToRemoveService = this.clickToRemoveService.bind(this);
     this.clickToUnblock = this.clickToUnblock.bind(this);
     this.showParticipant = this.showParticipant.bind(this);
 
-    this.elementId = elementId;
-    this.userRepository = userRepository;
-    this.conversationRepository = conversationRepository;
-    this.integrationRepository = integrationRepository;
-    this.teamRepository = teamRepository;
-    this.logger = new z.util.Logger('z.ViewModel.ParticipantsViewModel', z.config.LOGGER.OPTIONS);
+    this.elementId = 'participants';
+    this.mainViewModel = mainViewModel;
+
+    this.userRepository = repositories.user;
+    this.conversationRepository = repositories.conversation;
+    this.integrationRepository = repositories.integration;
+    this.teamRepository = repositories.team;
+    this.logger = new z.util.Logger('z.viewModel.details.ParticipantsViewModel', z.config.LOGGER.OPTIONS);
 
     this.state = ko.observable(ParticipantsViewModel.STATE.PARTICIPANTS);
     this.previousState = ko.observable(this.state());
     this.state.subscribe(oldState => this.previousState(oldState), null, 'beforeChange');
 
     this.activeAddState = ko.pureComputed(() => ParticipantsViewModel.CONFIG.ADD_STATES.includes(this.state()));
-    this.activeServiceState = ko.pureComputed(() => ParticipantsViewModel.CONFIG.SERVICE_STATES.includes(this.state()));
-
-    this.stateAddPeople = ko.pureComputed(() => this.state() === z.ViewModel.ParticipantsViewModel.STATE.ADD_PEOPLE);
-    this.stateAddService = ko.pureComputed(() => this.state() === z.ViewModel.ParticipantsViewModel.STATE.ADD_SERVICE);
-    this.stateParticipants = ko.pureComputed(() => {
-      return this.state() === z.ViewModel.ParticipantsViewModel.STATE.PARTICIPANTS;
+    this.activeServiceState = ko.pureComputed(() => {
+      return ParticipantsViewModel.CONFIG.SERVICE_STATES.includes(this.state());
     });
+
+    this.stateAddPeople = ko.pureComputed(() => this.state() === ParticipantsViewModel.STATE.ADD_PEOPLE);
+    this.stateAddService = ko.pureComputed(() => this.state() === ParticipantsViewModel.STATE.ADD_SERVICE);
+    this.stateParticipants = ko.pureComputed(() => this.state() === ParticipantsViewModel.STATE.PARTICIPANTS);
     this.stateServiceConfirmation = ko.pureComputed(() => {
-      return this.state() === z.ViewModel.ParticipantsViewModel.STATE.SERVICE_CONFIRMATION;
+      return this.state() === ParticipantsViewModel.STATE.SERVICE_CONFIRMATION;
     });
     this.stateServiceDetails = ko.pureComputed(() => {
-      return this.state() === z.ViewModel.ParticipantsViewModel.STATE.SERVICE_DETAILS;
+      return this.state() === ParticipantsViewModel.STATE.SERVICE_DETAILS;
     });
 
     this.conversation = ko.observable(new z.entity.Conversation());
@@ -194,7 +196,6 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
   }
 
   changeConversation(conversationEntity) {
-    this.participantsBubble.hide();
     this.conversation(conversationEntity);
     this.resetView();
   }
@@ -206,7 +207,6 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     }
 
     amplify.publish(z.event.WebApp.CONVERSATION.CREATE_GROUP, 'conversation_details', this.selectedParticipant());
-    this.participantsBubble.hide();
   }
 
   clickOnAddService() {
@@ -215,11 +215,9 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
   }
 
   clickOnPending(userEntity) {
-    const onSuccess = () => this.participantsBubble.hide();
-
     this.confirmDialog = $('#participants').confirm({
-      cancel: () => this.userRepository.ignore_connection_request(userEntity).then(() => onSuccess()),
-      confirm: () => this.userRepository.accept_connection_request(userEntity, true).then(() => onSuccess()),
+      cancel: () => this.userRepository.ignore_connection_request(userEntity),
+      confirm: () => this.userRepository.accept_connection_request(userEntity, true),
       data: {
         user: this.selectedParticipant(),
       },
@@ -262,12 +260,10 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     if (this.conversation().is_group()) {
       this.conversationRepository.addMembers(this.conversation(), this.selectedContacts());
     }
-    this.participantsBubble.hide();
   }
 
   clickToAddService() {
     this.integrationRepository.addService(this.conversation(), this.selectedService(), 'conversation_details');
-    this.participantsBubble.hide();
   }
 
   clickToBlock(userEntity) {
@@ -276,8 +272,6 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     this.confirmDialog = $('#participants').confirm({
       confirm: () => {
         const nextConversationEntity = this.conversationRepository.get_next_conversation(this.conversation());
-
-        this.participantsBubble.hide();
         this.userRepository.block_user(userEntity, nextConversationEntity);
       },
       data: {
@@ -285,10 +279,6 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
       },
       template: '#template-confirm-block',
     });
-  }
-
-  clickToConnect(userEntity) {
-    this.participantsBubble.hide();
   }
 
   clickToEdit() {
@@ -301,10 +291,7 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
     amplify.publish(z.event.WebApp.AUDIO.PLAY, z.audio.AudioType.ALERT);
 
     this.confirmDialog = $('#participants').confirm({
-      confirm: () => {
-        this.participantsBubble.hide();
-        this.conversationRepository.removeMember(this.conversation(), this.userRepository.self().id);
-      },
+      confirm: () => this.conversationRepository.removeMember(this.conversation(), this.userRepository.self().id),
       template: '#template-confirm-leave',
     });
   }
@@ -329,11 +316,8 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
 
   clickToRemoveService() {
     this.integrationRepository.removeService(this.conversation(), this.selectedParticipant()).then(response => {
-      if (response) {
-        if (this.groupMode()) {
-          return this.resetView();
-        }
-        this.participantsBubble.hide();
+      if (response && this.groupMode()) {
+        this.resetView();
       }
     });
   }
@@ -343,10 +327,7 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
       confirm: () => {
         this.userRepository
           .unblock_user(userEntity)
-          .then(() => {
-            this.participantsBubble.hide();
-            return this.conversationRepository.get_1to1_conversation(userEntity);
-          })
+          .then(() => this.conversationRepository.get_1to1_conversation(userEntity))
           .then(conversationEntity => this.conversationRepository.update_participating_user_ets(conversationEntity));
       },
       data: {
@@ -393,6 +374,10 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
       }
       this.selectedParticipant(participantEntity);
     }
+
+    if (!this.mainViewModel.isPanelOpen()) {
+      this.mainViewModel.openPanel();
+    }
   }
 
   showService(userEntity) {
@@ -410,76 +395,37 @@ z.ViewModel.ParticipantsViewModel = class ParticipantsViewModel {
   }
 
   switchContent(contentState) {
-    const isConnectionRequests = contentState === z.ViewModel.content.CONTENT_STATE.CONNECTION_REQUESTS;
+    const isConnectionRequests = contentState === z.viewModel.ContentViewModel.STATE.CONNECTION_REQUESTS;
     if (isConnectionRequests) {
-      this.participantsBubble.hide();
+      // todo: what to do?
     }
   }
 
-  toggleParticipantsBubble(addPeople = false) {
-    const toggleBubble = () => {
-      if (!this.participantsBubble.is_visible()) {
-        this.resetView();
-
-        const [userEntity] = this.participants();
-        const initialUser = userEntity && this.conversation().is_one2one() ? userEntity : undefined;
-        this.selectedParticipant(initialUser);
-
-        this.renderParticipants(true);
-      }
-
-      if (addPeople && !this.conversation().is_guest()) {
-        if (!this.participantsBubble.is_visible()) {
-          return this.participantsBubble.show().then(() => this.clickOnAddPeople());
-        }
-
-        const isConfirmingAction = this.confirmDialog && this.confirmDialog.is_visible();
-        if (this.stateAddPeople() || isConfirmingAction) {
-          return this.participantsBubble.hide();
-        }
-
-        return this.clickOnAddPeople();
-      }
-
-      return this.participantsBubble.toggle();
-    };
-
-    const bubble = wire.app.view.content.message_list.participant_bubble;
-    const timeout = bubble && bubble.is_visible() ? z.motion.MotionDuration.LONG : 0;
-    window.setTimeout(() => toggleBubble(), timeout);
-  }
-
   toggleParticipantsSidebar(addPeople = false) {
-    const mainView = wire.app.view.main;
-    const isOpen = mainView.isPanelOpen();
-    const toggleSidebar = () => {
-      if (!isOpen) {
-        this.resetView();
+    if (!this.mainViewModel.isPanelOpen()) {
+      this.resetView();
 
-        const [userEntity] = this.participants();
-        const initialUser = userEntity && this.conversation().is_one2one() ? userEntity : undefined;
-        this.selectedParticipant(initialUser);
+      const [userEntity] = this.participants();
+      const initialUser = userEntity && this.conversation().is_one2one() ? userEntity : undefined;
+      this.selectedParticipant(initialUser);
 
-        this.renderParticipants(true);
-      }
+      this.renderParticipants(true);
+    }
 
-      if (addPeople && !this.conversation().is_guest()) {
-        if (!isOpen) {
-          mainView.openPanel();
-          this.clickOnAddPeople();
-          return;
-        }
-
-        const isConfirmingAction = this.confirmDialog && this.confirmDialog.is_visible();
-        if (this.stateAddPeople() || isConfirmingAction) {
-          return mainView.closeRightColumn();
-        }
-
+    if (addPeople && !this.conversation().is_guest()) {
+      if (!this.mainViewModel.isPanelOpen()) {
+        this.mainViewModel.openPanel();
         return this.clickOnAddPeople();
       }
 
-      return mainView.togglePanel();
-    };
-    toggleSidebar();
+      const isConfirmingAction = this.confirmDialog && this.confirmDialog.is_visible();
+      if (this.stateAddPeople() || isConfirmingAction) {
+        return this.mainViewModel.closePanel();
+      }
+
+      return this.clickOnAddPeople();
+    }
+
+    this.mainViewModel.togglePanel();
   }
 };
